castingline_backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
psycopg2-binary==2.9.6
pandas==2.0.2
numpy==1.24.2
openpyxl==3.1.2
xlrd==2.0.1
gunicorn
